+++ pg_mentat/src/functions/store_management.rs
@@ -208,7 +208,7 @@ pub fn create_store(
 
         INSERT INTO {schema}.partitions (name, start_entid, end_entid, next_entid, allow_excision) VALUES
             ('db.part/db',   0,             1000000,       100,            FALSE),
-            ('db.part/user', 1000000,       1000000000000, 1000000,        FALSE),
+            ('db.part/user', 1000001,       1000000000000, 1000001,        FALSE),
             ('db.part/tx',   1000000000000, 2000000000000, 1000000000001,  FALSE)
         ON CONFLICT (name) DO NOTHING;
 
@@ -225,7 +225,7 @@ pub fn create_store(
         CREATE SEQUENCE IF NOT EXISTS {schema}.partition_db_seq
             START WITH 100 MINVALUE 100 MAXVALUE 999999 CACHE 10;
         CREATE SEQUENCE IF NOT EXISTS {schema}.partition_user_seq
-            START WITH 1000000 MINVALUE 1000000 MAXVALUE 999999999999 CACHE 100;
+            START WITH 1000001 MINVALUE 1000001 MAXVALUE 999999999999 CACHE 100;
         CREATE SEQUENCE IF NOT EXISTS {schema}.partition_tx_seq
             START WITH 1000000000001 MINVALUE 1000000000000 MAXVALUE 1999999999999 CACHE 100;
         ",